third_party/blink/web_tests/external/wpt/html/semantics/popovers/popover-focus-2_test3-expected.txt

This is a testharness.js-based test.
[FAIL] Popover focus navigation with imperative invocation
  assert_equals: Focus should move to popover0 since it is focusable expected Element node <div popover="" id="popover0" tabindex="0" style="top:300... but got Element node <button id="invoker1" tabindex="0">Invoker1</button>
Harness: the test ran to completion.
